+++ actionpack/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Rails 4.0.0 (unreleased) ##
 
+*   `truncate` now always returns an escaped HTMl-safe string. The option `:escape` can be used as
+    false to not escape the result.
+
+    *Li Ellis Gallardo + Rafael Mendonça França*
+
 *   `truncate` now accepts a block to show extra content when the text is truncated. *Li Ellis Gallardo*
 
 *   Add `week_field`, `week_field_tag`, `month_field`, `month_field_tag`, `datetime_local_field`,